pyproject.toml
[project]
name = "jira-ticket-exporter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.0",
    "types-requests>=2.32.0",
]
